+++ resources/assets/js/pages/administration/users/Show.vue
@@ -1,235 +1,118 @@
 <template>
 
     <div v-if="profile">
-        <div class="box profile-heading">
-            <div class="columns is-multiline">
-                <div class="column is-half-desktop">
-                    <div class="columns is-mobile">
-                        <div class="column is-narrow">
-                            <div class="image is-128x128">
-                                <img class="is-rounded"
-                                    :src="avatarLink">
-                            </div>
-                        </div>
-                        <div class="column">
-                            <p class="title is-4 has-padding-top-medium">
+        <div class="profile-heading">
+            <div class="columns is-centered">
+                <div class="column is-two-thirds-desktop">
+                    <div class="has-padding-medium">
+                        <div class="box is-raised">
+                            <p class="title is-4 has-text-centered">
                                 {{ profile.fullName }}
                             </p>
-                            <p class="subtitle is-5">
-                                {{ __('role') }}: {{ profile.role.name }}
-                            </p>
-                            <div class="user-controls has-margin-top-small"
-                                v-if="isSelfVisiting">
-                                <button class="button is-warning"
-                                    v-if="avatarId"
-                                    @click="deleteAvatar">
-                                    <span class="icon">
-                                        <fa icon="sync-alt"/>
-                                    </span>
-                                    <span v-if="!isTouch">
-                                        {{ __('Avatar') }}
-                                    </span>
-                                </button>
-                                <file-uploader @upload-successful="setUserAvatar($event.id)"
-                                    :url="uploadAvatarLink"
-                                    file-key="avatar">
-                                    <template slot="upload-button"
-                                        slot-scope="{ openFileBrowser }">
-                                        <button class="button is-info"
-                                            @click="openFileBrowser">
-                                            <span class="icon">
-                                                <fa icon="upload"/>
-                                            </span>
-                                            <span v-if="!isTouch">
-                                                {{ __('Avatar') }}
-                                            </span>
-                                        </button>
-                                    </template>
-                                </file-uploader>
-                                <button class="button is-danger is-pulled-right"
-                                    @click="exit()">
-                                    <span class="icon">
-                                        <fa icon="sign-out-alt"/>
-                                    </span>
-                                    <span v-if="!isTouch">
-                                        {{ __('Log Out') }}
-                                    </span>
-                                </button>
+                            <hr>
+                            <div class="columns">
+                                <div class="column">
+                                    <div class="columns is-mobile">
+                                        <div class="column">
+                                            <figure class="image is-128x128 avatar">
+                                                <img class="is-rounded is-raised"
+                                                    :src="avatarLink">
+                                            </figure>
+                                        </div>
+                                        <div class="column">
+                                            <div class="has-margin-top-small field controls"
+                                                v-if="isSelfVisiting">
+                                                <button class="button is-fullwidth is-warning"
+                                                    v-if="avatarId"
+                                                    @click="deleteAvatar">
+                                                    <span class="icon">
+                                                        <fa icon="sync-alt"/>
+                                                    </span>
+                                                    <span>
+                                                        {{ __('Avatar') }}
+                                                    </span>
+                                                </button>
+                                                <file-uploader @upload-successful="setUserAvatar($event.id)"
+                                                    :url="uploadAvatarLink"
+                                                    file-key="avatar">
+                                                    <template slot="upload-button"
+                                                        slot-scope="{ openFileBrowser }">
+                                                        <button class="button is-fullwidth is-info has-margin-top-small"
+                                                            @click="openFileBrowser">
+                                                            <span class="icon">
+                                                                <fa icon="upload"/>
+                                                            </span>
+                                                            <span>
+                                                                {{ __('Avatar') }}
+                                                            </span>
+                                                        </button>
+                                                    </template>
+                                                </file-uploader>
+                                                <button class="button is-fullwidth is-danger has-margin-top-small"
+                                                    @click="exit()">
+                                                    <span class="icon">
+                                                        <fa icon="sign-out-alt"/>
+                                                    </span>
+                                                    <span>
+                                                        {{ __('Log out') }}
+                                                    </span>
+                                                </button>
+                                            </div>
+                                            <div class="has-margin-top-large"
+                                                v-else>
+                                                <button class="button is-fullwidth is-warning"
+                                                    @click="$bus.$emit('start-impersonating', profile.id)"
+                                                    v-if="
+                                                        canAccess('core.impersonate.start')
+                                                        && !impersonating
+                                                    ">
+                                                    {{ __('Impersonate') }}
+                                                </button>
+                                            </div>
+                                        </div>
+                                    </div>
+                                </div>
+                                <div class="column">
+                                    <hr v-if="isMobile">
+                                    <div class="columns is-mobile">
+                                        <div class="column is-one-third has-text-right">
+                                            <p class="has-margin-top-small"><strong>{{ __('Owner') }}</strong></p>
+                                            <p class="has-margin-top-small"><strong>{{ __('Role') }}</strong></p>
+                                            <p class="has-margin-top-small"><strong>{{ __('Email') }}</strong></p>
+                                            <p class="has-margin-top-small"><strong>{{ __('Phone') }}</strong></p>
+                                        </div>
+                                        <div class="column">
+                                            <p class="has-margin-top-small">{{ profile.owner.name }}</p>
+                                            <p class="has-margin-top-small">{{ profile.role.name }}</p>
+                                            <p class="has-margin-top-small">{{ profile.email }}</p>
+                                            <p class="has-margin-top-small">{{ profile.phone }}</p>
+                                        </div>
+                                    </div>
+                                </div>
                             </div>
-                            <div class="has-margin-top-small"
-                                v-else>
-                                <button class="button is-warning"
-                                    @click="$bus.$emit('start-impersonating', profile.id)"
-                                    v-if="
-                                        canAccess('core.impersonate.start')
-                                        && !impersonating
-                                    ">
-                                    {{ __('Impersonate') }}
-                                </button>
+                            <hr>
+                            <div class="level is-mobile has-margin-top-medium">
+                                <div class="level-item has-text-centered has-right-border">
+                                    <div>
+                                        <p class="stat-value">{{ profile.loginCount }}</p>
+                                        <p class="stat-key">{{ __('logins') }}</p>
+                                    </div>
+                                </div>
+                                <div class="level-item has-text-centered">
+                                    <div>
+                                        <p class="stat-value">{{ profile.actionLogCount }}</p>
+                                        <p class="stat-key">{{ __('actions') }}</p>
+                                    </div>
+                                </div>
+                                <div class="level-item has-text-centered has-left-border">
+                                    <div>
+                                        <p class="stat-value">{{ profile.rating }}</p>
+                                        <p class="stat-key">{{ __('rating') }}</p>
+                                    </div>
+                                </div>
                             </div>
                         </div>
                     </div>
-                </div>
-                <div class="column is-half-desktop">
-                    <div class="level is-mobile has-margin-top-medium">
-                        <div class="level-item has-text-centered">
-                            <div>
-                                <p class="stat-value">{{ profile.loginCount }}</p>
-                                <p class="stat-key">{{ __('logins') }}</p>
-                            </div>
-                        </div>
-                        <div class="level-item has-text-centered has-lateral-borders">
-                            <div>
-                                <p class="stat-value">{{ profile.actionLogCount }}</p>
-                                <p class="stat-key">{{ __('actions') }}</p>
-                            </div>
-                        </div>
-                        <div class="level-item has-text-centered">
-                            <div>
-                                <p class="stat-value">{{ profile.rating }}</p>
-                                <p class="stat-key">{{ __('rating') }}</p>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="columns is-centered">
-            <div class="column is-two-thirds-tablet">
-                <div class="box">
-                    <nav class="pagination is-small has-margin-bottom-large">
-                        <a class="pagination-previous"
-                            @click="getPage(profile.timeline.current_page - 1)"
-                            :disabled="profile.timeline.prev_page_url===null">
-                            {{ __('Previous') }}
-                        </a>
-                        <a class="pagination-next"
-                            @click="getPage(profile.timeline.current_page + 1)"
-                            :disabled="profile.timeline.next_page_url===null">
-                            {{ __('Next') }}
-                        </a>
-                        <ul class="pagination-list" v-if="isShort">
-                            <li v-for="i in profile.timeline.last_page"
-                                :key="i">
-                                <a class="pagination-link"
-                                    :class="{ 'is-current': profile.timeline.current_page === i}"
-                                    @click="getPage(i)">
-                                    {{ i }}
-                                </a>
-                            </li>
-                        </ul>
-                        <ul class="pagination-list" v-if="isAtStart && !isShort">
-                            <li v-for="i in 3"
-                                :key="i">
-                                <a class="pagination-link"
-                                    :class="{ 'is-current': profile.timeline.current_page === i}"
-                                    @click="getPage(i)">
-                                    {{ i }}
-                                </a>
-                            </li>
-                            <li><span class="pagination-ellipsis">&hellip;</span></li>
-                            <li>
-                                <a class="pagination-link"
-                                    @click="getPage(profile.timeline.last_page)">
-                                    {{ profile.timeline.last_page }}
-                                </a>
-                            </li>
-                        </ul>
-                        <ul class="pagination-list" v-if="hasMiddle && !isShort">
-                            <li>
-                                <a class="pagination-link"
-                                    @click="getPage(1)">
-                                    1
-                                </a>
-                            </li>
-                            <li><span class="pagination-ellipsis">&hellip;</span></li>
-                            <li>
-                                <a class="pagination-link"
-                                    @click="getPage(profile.timeline.current_page - 1)">
-                                    {{ profile.timeline.current_page - 1 }}
-                                </a>
-                            </li>
-                            <li>
-                                <a class="pagination-link is-current">
-                                    {{ profile.timeline.current_page }}
-                                </a>
-                            </li>
-                            <li>
-                                <a class="pagination-link"
-                                    @click="getPage(profile.timeline.current_page + 1)">
-                                    {{ profile.timeline.current_page + 1 }}
-                                </a>
-                            </li>
-                            <li><span class="pagination-ellipsis">&hellip;</span></li>
-                            <li>
-                                <a class="pagination-link"
-                                    @click="getPage(profile.timeline.last_page)">
-                                    {{ profile.timeline.last_page }}
-                                </a>
-                            </li>
-                        </ul>
-                        <ul class="pagination-list" v-if="isAtEnd && !isShort">
-                            <li>
-                                <a class="pagination-link"
-                                    @click="getPage(1)">
-                                    1
-                                </a>
-                            </li>
-                            <li><span class="pagination-ellipsis">&hellip;</span></li>
-                            <li v-for="i in 3"
-                                :key="i">
-                                <a class="pagination-link"
-                                    :class="{
-                                        'is-current': profile.timeline.current_page
-                                            === profile.timeline.last_page - 3 + i
-                                    }"
-                                    @click="getPage(profile.timeline.last_page - 3 + i)">
-                                    {{ profile.timeline.last_page - 3 + i }}
-                                </a>
-                            </li>
-                        </ul>
-                    </nav>
-                    <ul class="timeline has-padding-large">
-                        <div v-for="(actions, day) in timeline"
-                            class="timeline-content"
-                            :key="day">
-                            <li class="timeline-header">
-                                <span class="tag is-medium is-primary">{{ __(day) }}</span>
-                            </li>
-                            <!-- <li class="timeline-item">
-                                <div class="timeline-marker"></div>
-                                <div class="timeline-content">
-                                    <p class="heading">January 2016</p>
-                                    <p>Timeline content - Can include any HTML element</p>
-                                </div>
-                            </li> -->
-                            <li class="timeline-item"
-                                v-for="(action, index) in actions"
-                                :key="index">
-                                <div class="timeline-marker is-icon"
-                                    :class="getClass(action.route)">
-                                    <fa :icon="getIcon(action.route)" size="xs"/>
-                                </div>
-                                <div class="timeline-content">
-                                    <p class="heading">
-                                        {{ getHRDate(action.created_at) }}
-                                        {{ getHRTime(action.created_at) }}
-                                    </p>
-                                    <p>{{ action.permission.description }}</p>
-                                </div>
-                            </li>
-                        </div>
-                        <li class="timeline-header"
-                            v-if="profile.timeline.current_page === profile.timeline.last_page">
-                            <span class="tag is-medium is-primary">{{ __('End') }}</span>
-                        </li>
-                        <li class="timeline-item"
-                            v-else>
-                            <div class="timeline-marker is-icon">
-                                <fa icon="ellipsis-h" size="xs"/>
-                            </div>
-                        </li>
-                    </ul>
                 </div>
             </div>
         </div>
@@ -241,17 +124,10 @@
 
 import { mapState, mapMutations } from 'vuex';
 import { library } from '@fortawesome/fontawesome-svg-core';
-import {
-    faSyncAlt, faTrashAlt, faUpload, faSignOutAlt,
-    faEllipsisH, faEye, faPlus, faPencilAlt,
-} from '@fortawesome/free-solid-svg-icons';
+import { faSyncAlt, faTrashAlt, faUpload, faSignOutAlt } from '@fortawesome/free-solid-svg-icons';
 import FileUploader from '../../../components/enso/fileuploader/FileUploader.vue';
-import format from '../../../modules/enso/plugins/date-fns/format';
 
-library.add([
-    faSyncAlt, faTrashAlt, faUpload, faSignOutAlt,
-    faEllipsisH, faEye, faPlus, faPencilAlt,
-]);
+library.add(faSyncAlt, faTrashAlt, faUpload, faSignOutAlt);
 
 export default {
     components: { FileUploader },
@@ -265,7 +141,7 @@ export default {
     computed: {
         ...mapState(['user', 'meta', 'impersonating']),
         ...mapState('auth', ['isAuth']),
-        ...mapState('layout', ['isTouch']),
+        ...mapState('layout', ['isMobile']),
         uploadAvatarLink() {
             return route('core.avatars.store');
         },
@@ -280,43 +156,6 @@ export default {
         avatarLink() {
             return route('core.avatars.show', this.avatarId);
         },
-        timeline() {
-            const actions = this.profile.timeline.data;
-
-            return actions.map(action => action.created_at).reduce((days, record) => {
-                days.push(this.getDay(record));
-                return days;
-            }, []).filter((value, index, days) => days.indexOf(value) === index)
-                .reduce((timeline, day) => {
-                    timeline[day] = actions
-                        .filter(record => this.getDay(record.created_at) === day);
-                    return timeline;
-                }, {});
-        },
-        isShort() {
-            return this.profile.timeline.last_page <= 5;
-        },
-        isAtStart() {
-            return this.profile.timeline.current_page <= 2;
-        },
-        isAtEnd() {
-            return this.profile.timeline.last_page
-                - this.profile.timeline.current_page <= 2;
-        },
-        hasMiddle() {
-            return !this.isAtStart && !this.isAtEnd;
-        },
-        paginationUrl() {
-            return `${route('administration.users.show', this.profile.id)}?page=`;
-        },
-    },
-
-    watch: {
-        locale: {
-            handler() {
-                this.getPage(this.profile.timeline.current_page);
-            },
-        },
     },
 
     created() {
@@ -329,9 +168,8 @@ export default {
         ...mapMutations(['setUserAvatar', 'initialise']),
         getProfile() {
             axios.get(route(this.$route.name, this.$route.params.id))
-                .then((response) => {
-                    this.profile = response.data.user;
-                }).catch(error => this.handleError(error));
+                .then(response => (this.profile = response.data.user))
+                .catch(error => this.handleError(error));
         },
         deleteAvatar() {
             axios.delete(route('core.avatars.destroy', this.user.avatarId))
@@ -348,41 +186,26 @@ export default {
                 this.handleError(error);
             });
         },
-        getDay(timestamp) {
-            return format(timestamp, 'dddd');
-        },
-        getHRDate(timestamp) {
-            return format(timestamp, this.meta.dateFormat);
-        },
-        getHRTime(timestamp) {
-            return format(timestamp, 'H:mm');
-        },
-        getIcon(action) {
-            if (action.indexOf('index') > 0) return faEye;
-            if (action.indexOf('create') > 0) return faPlus;
-            if (action.indexOf('edit') > 0) return faPencilAlt;
-            return faTrashAlt;
-        },
-        getClass(action) {
-            if (action.indexOf('index') > 0) return 'has-text-success';
-            if (action.indexOf('create') > 0) return 'has-text-warning';
-            if (action.indexOf('edit') > 0) return 'has-text-warning';
-            return 'has-text-danger';
-        },
-        getPage(page) {
-            axios.get(this.paginationUrl + page).then(({ data }) => {
-                this.profile = data.user;
-            }).catch(error => this.handleError(error));
-        },
     },
 };
 
 </script>
 
-<style>
+<style lang="scss" scoped>
 
-    .has-lateral-borders {
+    .avatar {
+        margin: auto;
+    }
+
+    .controls, .details {
+        justify-content: center;
+    }
+
+    .has-left-border {
         border-left: 1px solid rgba(0,0,0,0.2);
+    }
+
+    .has-right-border {
         border-right: 1px solid rgba(0,0,0,0.2);
     }
 
@@ -395,14 +218,6 @@ export default {
         font-size: 1.4em;
         font-weight: 200;
         padding-bottom: 8px;
-    }
-
-    .level.user-controls {
-        margin-bottom: 0;
-    }
-
-    .timeline-content {
-        transition:all 1s ease;
     }
 
 </style>